% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubicSpline-package.R
\docType{package}
\name{cubicSpline-package}
\alias{cubicSpline}
\alias{cubicSpline-package}
\title{cubicSpline: Cubic Spline Fitting}
\description{
Implements B-spline curve fitting using Cox-de Boor recursion. Provides tools for fitting and evaluating B-spline curves.
}
\keyword{internal}
