% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline_basis.R
\name{bspline_basis}
\alias{bspline_basis}
\title{B-spline Basis Function}
\usage{
bspline_basis(x, i, order = 4, knots)
}
\arguments{
\item{x}{Numeric vector at which to evaluate the basis function}

\item{i}{Integer index of the basis function}

\item{order}{Integer order of the B-spline, default 4}

\item{knots}{Numeric vector of knot positions}
}
\value{
A numeric value representing how much this particular piece of the B-spline curve contributes at point x.
}
\description{
B-spline Basis Function
}
\examples{
# Create knots and evaluation points
knots <- c(0, 0, 0, 0, 1, 2, 3, 4, 4, 4, 4)
x <- seq(0, 4, length.out = 100)
basis1 <- sapply(x, bspline_basis, i = 1, order = 3, knots = knots)
print(basis1)
}
