% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_bspline.R
\name{bspline_basis}
\alias{bspline_basis}
\alias{my_bspline}
\title{Evaluate a B-spline basis function}
\usage{
my_bspline(x0, y0)
}
\arguments{
\item{x}{A numeric value where the B-spline is evaluated.}

\item{i}{The index of the basis function.}

\item{order}{The order of the B-spline.}

\item{knots}{A numeric vector of knots for the B-spline.}
}
\value{
A numeric value representing the evaluated B-spline basis function.
}
\description{
This function evaluates a B-spline basis function using Cox-de Boor recursion.
}
